chore: light refactoring of v2x demo client
